+++ massiv/src/Data/Massiv/Array/Manifest/Storable.hs
@@ -39,6 +39,7 @@ import Data.Massiv.Array.Manifest.Internal
 import Data.Massiv.Array.Manifest.Primitive (shrinkMutableByteArray)
 import Data.Primitive.ByteArray (MutableByteArray(..))
 import Data.Massiv.Array.Manifest.List as A
+import Data.Massiv.Array.Manifest.Vector.Stream as S (steps)
 import Data.Massiv.Array.Mutable
 import Data.Massiv.Core.Common
 import Data.Massiv.Core.List
@@ -208,6 +209,10 @@ instance (Index ix, VS.Storable e) => Load S ix e where
   {-# INLINE loadArrayM #-}
 
 instance (Index ix, VS.Storable e) => StrideLoad S ix e
+
+instance (Index ix, VS.Storable e) => Stream S ix e where
+  toStream = S.steps
+  {-# INLINE toStream #-}
 
 
 instance ( VS.Storable e