+++ massiv/src/Data/Massiv/Array/Ops/Construct.hs
@@ -30,6 +30,8 @@ module Data.Massiv.Array.Ops.Construct
   , iterateN
   , iiterateN
     -- *** Unfolding
+  , unfoldr
+  , unfoldrN
   , unfoldlS_
   , iunfoldlS_
   , unfoldrS_
@@ -67,6 +69,7 @@ import Control.Monad (when, void)
 import Control.Monad.ST
 import Data.Massiv.Array.Delayed.Pull
 import Data.Massiv.Array.Delayed.Push
+import Data.Massiv.Array.Delayed.Stream (unfoldr, unfoldrN)
 import Data.Massiv.Array.Mutable
 import Data.Massiv.Core.Common
 import Prelude as P hiding (enumFromTo, replicate)